# Vectorize `node_type`