--- a/stellargraph/data/explorer.py
+++ b/stellargraph/data/explorer.py
@@ -571,12 +571,13 @@ class UniformRandomMetaPathWalk(RandomWalk):
                         # d+1 can also be used to index metapath to retrieve the node type for the next step in the walk
                         neighbours = self.graph.neighbors(current_node, use_ilocs=True)
                         # filter these by node type
+                        neighbour_types = self.graph.vectorized_node_type(neighbours)
                         neighbours = [
-                            n_node
-                            for n_node in neighbours
-                            if self.graph.node_type(n_node, use_ilocs=True)
-                            == metapath[d]
+                            neigh
+                            for neigh, neigh_type in zip(neighbours, neighbour_types)
+                            if neigh_type == metapath[d]
                         ]
+
                         if len(neighbours) == 0:
                             # if no neighbours of the required type as dictated by the metapath exist, then stop.
                             break