@kieranricardo 2020-05-04
using `timeit -n 3 -r 5` yields:

- develop: 108 ms ± 1.86 ms
- this PR: 251 µs ± 26.6 µs